added a node-server to the autosave client that waits for code to save
The client will check every 3 secons and if the code didn't change since the last time it will save it (so it waits for a period of no changes). But it only saves the code once.
It does so by sending the code to the node backend process which takes care of the saving.
Files are saved in the folder of the node backend process and the filename will be the original filename of the source file preceded by a timestamp given by the client.